refactor(api): move template trigger and hook routes
X-Anneal-Run: cmthmesjt7vdumpxn4iqjnsos
X-Anneal-Step: 2: Implementation
X-Anneal-Provider: codex
